Changed SDK initialization
