+++ remo/config.py
@@ -1,9 +1,12 @@
 import json
 import os
+from pathlib import Path
+
+REMO_HOME = os.getenv('REMO_HOME', str(Path.home().joinpath('.remo')))
 
 
 class Config:
-    __slots__ = ['port', 'server', 'user_name', 'user_email', 'user_password', 'conda_env', 'viewer']
+    __slots__ = ['port', 'server', 'user_name', 'user_email', 'user_password', 'viewer']
     _default_port = 8123
     _default_server = 'http://localhost'
     _default_user_name = 'Admin User'
@@ -12,22 +15,18 @@ class Config:
     _default_viewer = 'browser'
 
     def __init__(self, config):
-        self.port = config.get('port', self._default_port)
-        self.server = config.get('server', self._default_server)
+        for name in self.__slots__:
+            setattr(self, name, config.get(name, getattr(self, '_default_{}'.format(name))))
 
-        self.user_name = config.get('user_name', self._default_user_name)
-        self.user_email = config.get('user_email', self._default_user_email)
-        self.user_password = config.get('user_password', self._default_user_password)
+    def server_url(self):
+        return '{}:{}'.format(self.server, self.port)
 
-        self.conda_env = config.get('conda_env')
-        self.viewer = config.get('viewer', self._default_viewer)
+    @staticmethod
+    def load(config_path: str = str(os.path.join(REMO_HOME, 'remo.json'))):
+        if not os.path.exists(config_path):
+            return None
 
+        with open(config_path) as cfg_file:
+            config = json.load(cfg_file)
 
-def parse_config(config_path):
-    if not os.path.exists(config_path):
-        return None
-
-    with open(config_path) as cfg_file:
-        config = json.load(cfg_file)
-
-    return Config(config)
+        return Config(config)
