+++ remo/domain/__init__.py
@@ -1,4 +1,4 @@
 from .annotation_set import AnnotationSet
 from .dataset import Dataset
 from .image import Image
-from .annotation import Annotation
+from .annotation import Annotation, Bbox, Segment
