+++ setup.py
@@ -23,7 +23,8 @@ setup(
         'Programming Language :: Python :: 3.5',
         'Programming Language :: Python :: 3.6',
         'Programming Language :: Python :: 3.7',
-        
+        'Programming Language :: Python :: 3.8',
+
         'Operating System :: OS Independent',
     ],
     python_requires='>=3.5',
